# Conflicts:
#	src/main/java/ru/example/group/main/config/SecurityConfig.java
#	src/main/java/ru/example/group/main/controller/UserRegisterController.java
#	src/main/java/ru/example/group/main/controller/UserSettingsController.java
#	src/main/java/ru/example/group/main/service/SearchService.java
#	src/main/java/ru/example/group/main/service/UserRegisterService.java
#	src/main/java/ru/example/group/main/service/UserSettingsService.java

+++ src/main/java/ru/example/group/main/service/UserRegisterService.java
@@ -181,7 +181,7 @@ public class UserRegisterService {
 
     private boolean localFileSizeEquallyRemoteFileSize() throws MalformedURLException, URISyntaxException {
         int sizeRemoteFile = getRemoteFileSize(new URL(remoteFileGeoLite2));
-        long sizeLocalFile = FileUtils.sizeOf(new File(Objects.requireNonNull(getClass().getClassLoader().getResource(localFileGeoLite2)).toURI()));
+        int sizeLocalFile = (int) FileUtils.sizeOf(new File(Objects.requireNonNull(getClass().getClassLoader().getResource(localFileGeoLite2)).toURI()));
         return sizeRemoteFile != sizeLocalFile;
     }
